# SchoolManagementClient/OpenApi/UserManagement.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "UserManagement",
    "version": "1.0"
  },
  "paths": {
...
    "/StudentRegister": {
      "post": {
        "tags": [
          "UserManagement"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentRegisterDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentRegisterDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StudentRegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/LecturerRegister": {
      "post": {
        "tags": [
          "UserManagement"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LecturerRegisterDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LecturerRegisterDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LecturerRegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "LecturerRegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "middleName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "identityNumber": {
            "type": "integer",
            "format": "int64"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date",
            "example": "2024-05-08"
          },
          "departmentId": {
            "type": "integer",
            "format": "int32"
          },
          "facultyId": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "nullable": true
          },
          "userType": {
            "$ref": "#/components/schemas/UserType"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "positionId": {
            "type": "integer",
            "format": "int32"
          },
          "employeeNumber": {
            "type": "string",
            "nullable": true
          },
          "hireDate": {
            "type": "string",
            "format": "date-time"
          },
          "lecturerStatus": {
            "$ref": "#/components/schemas/LecturerStatus"
          }
        },
        "additionalProperties": false
      },
      "LecturerStatus": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "StudentRegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "middleName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "identityNumber": {
            "type": "integer",
            "format": "int64"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date",
            "example": "2024-05-08"
          },
          "departmentId": {
            "type": "integer",
            "format": "int32"
          },
          "facultyId": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "nullable": true
          },
          "userType": {
            "$ref": "#/components/schemas/UserType"
          },
          "studentNumber": {
            "type": "string",
            "nullable": true
          },
          "enrollmentDate": {
            "type": "string",
            "format": "date-time"
          },
          "degreeProgram": {
            "type": "string",
            "nullable": true
          },
          "studentStatus": {
            "$ref": "#/components/schemas/StudentStatus"
          },
          "advisorId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "StudentStatus": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "UserType": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      }
    }
  }
}